+++ src/DataCleaner.Application/DependencyInjection.cs
@@ -1,5 +1,6 @@
 using DataCleaner.Application.Abstractions;
 using DataCleaner.Application.Import;
+using DataCleaner.Application.Profiling;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace DataCleaner.Application;
@@ -9,6 +10,7 @@ public static class DependencyInjection
     public static IServiceCollection AddApplication(this IServiceCollection services)
     {
         services.AddTransient<IDataImportService, DataImportService>();
+        services.AddTransient<IDataProfilingService, DataProfilingService>();
         return services;
     }
 }